change to u16

+++ src/uuid/service.rs
@@ -2,141 +2,141 @@
 //! 
 //! https://www.bluetooth.com/wp-content/uploads/Files/Specification/Assigned_Numbers.html#bookmark104
 
-pub const GAP: u8 = 0x1800;
+pub const GAP: u16 = 0x1800;
 
-pub const GATT: u8 = 0x1801;
+pub const GATT: u16 = 0x1801;
 
-pub const IMMEDIATE_ALERT: u8 = 0x1802;
+pub const IMMEDIATE_ALERT: u16 = 0x1802;
 
-pub const LINK_LOSS: u8 = 0x1803;
+pub const LINK_LOSS: u16 = 0x1803;
 
-pub const TX_POWER: u8 = 0x1804;
+pub const TX_POWER: u16 = 0x1804;
 
-pub const CURRENT_TIME: u8 = 0x1805;
+pub const CURRENT_TIME: u16 = 0x1805;
 
-pub const REFERENCE_TIME_UPDATE: u8 = 0x1806;
+pub const REFERENCE_TIME_UPDATE: u16 = 0x1806;
 
-pub const NEST_DST_CHANGE: u8 = 0x1807;
+pub const NEST_DST_CHANGE: u16 = 0x1807;
 
-pub const GLUCOSE: u8 = 0x1808;
+pub const GLUCOSE: u16 = 0x1808;
 
-pub const HEALTH_THERMOMETER: u8 = 0x1809;
+pub const HEALTH_THERMOMETER: u16 = 0x1809;
 
-pub const DEVICE_INFORMATION: u8 = 0x180A;
+pub const DEVICE_INFORMATION: u16 = 0x180A;
 
-pub const HEART_RATE: u8 = 0x180D;
+pub const HEART_RATE: u16 = 0x180D;
 
-pub const PHONE_ALERT_STATUS: u8 = 0x180E;
+pub const PHONE_ALERT_STATUS: u16 = 0x180E;
 
-pub const BATTERY: u8 = 0x180F;
+pub const BATTERY: u16 = 0x180F;
 
-pub const BLOOD_PRESSURE: u8 = 0x1810;
+pub const BLOOD_PRESSURE: u16 = 0x1810;
 
-pub const ALERT_NOTIFICATION: u8 = 0x1811;
+pub const ALERT_NOTIFICATION: u16 = 0x1811;
 
-pub const HUMAN_INTERFACE_DEVICE: u8 = 0x1812;
+pub const HUMAN_INTERFACE_DEVICE: u16 = 0x1812;
 
-pub const SCAN_PARAMETERS: u8 = 0x1813;
+pub const SCAN_PARAMETERS: u16 = 0x1813;
 
-pub const RUNNING_SPEED_AND_CADENCE: u8 = 0x1814;
+pub const RUNNING_SPEED_AND_CADENCE: u16 = 0x1814;
 
-pub const AUTOMATION_IO: u8 = 0x1815;
+pub const AUTOMATION_IO: u16 = 0x1815;
 
-pub const CYCLING_SPEED_AND_CADENCE: u8 = 0x1816;
+pub const CYCLING_SPEED_AND_CADENCE: u16 = 0x1816;
 
-pub const CYCLING_POWER: u8 = 0x1818;
+pub const CYCLING_POWER: u16 = 0x1818;
 
-pub const LOCATION_AND_NAVIGATION: u8 = 0x1819;
+pub const LOCATION_AND_NAVIGATION: u16 = 0x1819;
 
-pub const ENVIRONMENTAL_SENSING: u8 = 0x181A;
+pub const ENVIRONMENTAL_SENSING: u16 = 0x181A;
 
-pub const BODY_COMPOSITION: u8 = 0x181B;
+pub const BODY_COMPOSITION: u16 = 0x181B;
 
-pub const USER_DATA: u8 = 0x181C;
+pub const USER_DATA: u16 = 0x181C;
 
-pub const WEIGHT_SCALE: u8 = 0x181D;
+pub const WEIGHT_SCALE: u16 = 0x181D;
 
-pub const BOND_MANAGEMENT: u8 = 0x181E;
+pub const BOND_MANAGEMENT: u16 = 0x181E;
 
-pub const CONTINUOUS_GLUCOSE_MONITORING: u8 = 0x181F;
+pub const CONTINUOUS_GLUCOSE_MONITORING: u16 = 0x181F;
 
-pub const INTERNET_PROTOCOL_SUPPORT: u8 = 0x1820;
+pub const INTERNET_PROTOCOL_SUPPORT: u16 = 0x1820;
 
-pub const INDOOR_POSITIONING: u8 = 0x1821;
+pub const INDOOR_POSITIONING: u16 = 0x1821;
 
-pub const PULSE_OXIMETER: u8 = 0x1822;
+pub const PULSE_OXIMETER: u16 = 0x1822;
 
-pub const HTTP_PROXY: u8 = 0x1823;
+pub const HTTP_PROXY: u16 = 0x1823;
 
-pub const TRANSPORT_DISCOVERY: u8 = 0x1824;
+pub const TRANSPORT_DISCOVERY: u16 = 0x1824;
 
-pub const OBJECT_TRANSFER: u8 = 0x1825;
+pub const OBJECT_TRANSFER: u16 = 0x1825;
 
-pub const FITNESS_MACHINE: u8 = 0x1826;
+pub const FITNESS_MACHINE: u16 = 0x1826;
 
-pub const MESH_PROVISIONING: u8 = 0x1827;
+pub const MESH_PROVISIONING: u16 = 0x1827;
 
-pub const MESH_PROXY: u8 = 0x1828;
+pub const MESH_PROXY: u16 = 0x1828;
 
-pub const RECONNECTION_CONFIGURATION: u8 = 0x1829;
+pub const RECONNECTION_CONFIGURATION: u16 = 0x1829;
 
-pub const INSULIN_DELIVERY: u8 = 0x183A;
+pub const INSULIN_DELIVERY: u16 = 0x183A;
 
-pub const BINARY_SENSOR: u8 = 0x183B;
+pub const BINARY_SENSOR: u16 = 0x183B;
 
-pub const EMERGENCY_CONFIGURATION: u8 = 0x183C;
+pub const EMERGENCY_CONFIGURATION: u16 = 0x183C;
 
-pub const AUTHORIZATION_CONTROL: u8 = 0x183D;
+pub const AUTHORIZATION_CONTROL: u16 = 0x183D;
 
-pub const PHYSICAL_ACTIVITY_MONITOR: u8 = 0x183E;
+pub const PHYSICAL_ACTIVITY_MONITOR: u16 = 0x183E;
 
-pub const ELAPSED_TIME: u8 = 0x183F;
+pub const ELAPSED_TIME: u16 = 0x183F;
 
-pub const GENERIC_HEALTH_SENSOR: u8 = 0x1840;
+pub const GENERIC_HEALTH_SENSOR: u16 = 0x1840;
 
-pub const AUDIO_INPUT_CONTROL: u8 = 0x1843;
+pub const AUDIO_INPUT_CONTROL: u16 = 0x1843;
 
-pub const VOLUME_CONTROL: u8 = 0x1844;
+pub const VOLUME_CONTROL: u16 = 0x1844;
 
-pub const VOLUME_OFFSET_CONTROL: u8 = 0x1845;
+pub const VOLUME_OFFSET_CONTROL: u16 = 0x1845;
 
-pub const COORDINATED_SET_IDENTIFICATION: u8 = 0x1846;
+pub const COORDINATED_SET_IDENTIFICATION: u16 = 0x1846;
 
-pub const DEVICE_TIME: u8 = 0x1847;
+pub const DEVICE_TIME: u16 = 0x1847;
 
-pub const MEDIA_CONTROL: u8 = 0x1848;
+pub const MEDIA_CONTROL: u16 = 0x1848;
 
-pub const GENERIC_MEDIA_CONTROL: u8 = 0x1849;
+pub const GENERIC_MEDIA_CONTROL: u16 = 0x1849;
 
-pub const CONSTANT_TONE_EXTENSION: u8 = 0x184A;
+pub const CONSTANT_TONE_EXTENSION: u16 = 0x184A;
 
-pub const TELEPHONE_BEARER: u8 = 0x184B;
+pub const TELEPHONE_BEARER: u16 = 0x184B;
 
-pub const GENERIC_TELEPHONE_BEARER: u8 = 0x184C;
+pub const GENERIC_TELEPHONE_BEARER: u16 = 0x184C;
 
-pub const MICROPHONE_CONTROL: u8 = 0x184D;
+pub const MICROPHONE_CONTROL: u16 = 0x184D;
 
-pub const AUDIO_STREAM_CONTROL: u8 = 0x184E;
+pub const AUDIO_STREAM_CONTROL: u16 = 0x184E;
 
-pub const BROADCAST_AUDIO_SCAN: u8 = 0x184F;
+pub const BROADCAST_AUDIO_SCAN: u16 = 0x184F;
 
-pub const PUBLISHED_AUDIO_CAPABILITIES: u8 = 0x1850;
+pub const PUBLISHED_AUDIO_CAPABILITIES: u16 = 0x1850;
 
-pub const BASIC_AUDIO_ANNOUNCEMENT: u8 = 0x1851;
+pub const BASIC_AUDIO_ANNOUNCEMENT: u16 = 0x1851;
 
-pub const BROADCAST_AUDIO_ANNOUNCEMENT: u8 = 0x1852;
+pub const BROADCAST_AUDIO_ANNOUNCEMENT: u16 = 0x1852;
 
-pub const COMMON_AUDIO: u8 = 0x1853;
+pub const COMMON_AUDIO: u16 = 0x1853;
 
-pub const HEARING_ACCESS: u8 = 0x1854;
+pub const HEARING_ACCESS: u16 = 0x1854;
 
-pub const TELEPHONY_AND_MEDIA_AUDIO: u8 = 0x1855;
+pub const TELEPHONY_AND_MEDIA_AUDIO: u16 = 0x1855;
 
-pub const PUBLIC_BROADCAST_ANNOUNCEMENT: u8 = 0x1856;
+pub const PUBLIC_BROADCAST_ANNOUNCEMENT: u16 = 0x1856;
 
-pub const ELECTRONIC_SHELF_LABEL: u8 = 0x1857;
+pub const ELECTRONIC_SHELF_LABEL: u16 = 0x1857;
 
-pub const GAMING_AUDIO: u8 = 0x1858;
+pub const GAMING_AUDIO: u16 = 0x1858;
 
-pub const MESH_PROXY_SOLICITATION: u8 = 0x1859;
+pub const MESH_PROXY_SOLICITATION: u16 = 0x1859;
